backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.32.0",
    "alembic>=1.14.0",
    "pyjwt>=2.10.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.7.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "requests>=2.32.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests requiring external services (deselect with '-m \"not integration\"')",
]
